shared/requirements.txt
boto3
launchdarkly-server-sdk
pypdf
pypdf[crypto]
python-jose
reportlab
requests